requirements.txt
torch
torchvision
torchaudio
transformers
huggingface-hub
pillow
fire
tqdm
termcolor
wandb
peft
datasets
matplotlib
